requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
emoji==2.14.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.5
pandas==2.2.3
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
pyvi==0.1.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-crfsuite==0.5.0
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
